requirements.txt
aniso8601==1.2.0
click==6.6
cloudpickle==0.2.1
cymem==1.31.2
Flask==0.11.1
Flask-RESTful==0.3.5
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
murmurhash==0.26.4
numpy==1.11.2
pathlib==1.0.1
plac==0.9.6
preshed==0.46.4
python-dateutil==2.5.3
pytz==2016.7
semver==2.7.1
six==1.10.0
spacy==1.2.0
sputnik==0.9.3
thinc==5.0.8
ujson==1.35
Werkzeug==0.11.11
neo4j-driver==1.0.2
